requirements.txt
transformers
transformers[sentencepiece]
datasets
sacreBLEU
rouge_score
py7zr
pandas
numpy
nltk
PyYaml
matplotlib
torch
ensure
python-box
-e .